changed action and reducers also added reducer for loading state
