+++ app/actions/Types.js
@@ -6,7 +6,9 @@ export const ADD_INVENTORY = 'ADD_INVENTORY';
 export const UPDATE_PENDING_COUNT = 'UPDATE_PENDING_COUNT';
 export const UPDATE_UPLOAD_COUNT = 'UPDATE_UPLOAD_COUNT';
 export const IS_UPLOADING = 'IS_UPLOADING';
-export const SET_LOADING = 'SET_LOADING';
+
+export const UPDATE_LOADING = 'UPDATE_LOADING';
+
 export const SET_SIGNUP_LOADER = 'SET_SIGNUP_LOADER';
 export const SET_SPECIES_LIST = 'SET_SPECIES_LIST';
 export const SET_SPECIES_ID = 'SET_SPECIES_ID';
