+++ app/components/CreatePolygon/MapMarking.js
@@ -14,8 +14,7 @@ import { Header, PrimaryButton, Alrighty } from '../Common';
 import { Colors } from '_styles';
 import MapboxGL from '@react-native-mapbox-gl/maps';
 import { active_marker, marker_png } from '../../assets/index';
-import { addCoordinates, polygonUpdate } from '../../actions';
-import { getInventory } from '../../repositories/inventory';
+import { getInventory, addCoordinates, polygonUpdate } from '../../repositories/inventory';
 import { useNavigation } from '@react-navigation/native';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import Geolocation from '@react-native-community/geolocation';
