+++ app/components/RegisterSingleTree/MapMarking.js
@@ -15,8 +15,7 @@ import { Header, PrimaryButton, Alrighty } from '../Common';
 import { Colors } from '_styles';
 import MapboxGL from '@react-native-mapbox-gl/maps';
 import { active_marker, marker_png, off_site_enable_banner } from '../../assets';
-import { addCoordinateSingleRegisterTree } from '../../actions';
-import { getInventory } from '../../repositories/inventory';
+import { getInventory, addCoordinateSingleRegisterTree } from '../../repositories/inventory';
 import { useNavigation } from '@react-navigation/native';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 import Geolocation from '@react-native-community/geolocation';
