+++ app/reducers/provider.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { View, Text } from 'react-native';
+import { InventoryContextProvider } from './inventory';
+import { LoadingContextProvider } from './loader';
+
+export default function Provider({ children }) {
+  return (
+    <InventoryContextProvider>
+      <LoadingContextProvider>{children}</LoadingContextProvider>
+    </InventoryContextProvider>
+  );
+}
