+++ app/repositories/inventory.js
@@ -9,7 +9,7 @@ import {
   AddSpecies,
 } from '../actions/Schemas';
 import { bugsnag } from '../utils';
-import { LocalInventoryActions } from '../actions/Action';
+import { updateCount } from '../actions/inventory';
 
 export const getInventoryByStatus = (status) => {
   return new Promise((resolve, reject) => {
@@ -93,10 +93,10 @@ export const changeInventoryStatusAndResponse = ({ inventory_id, status, respons
             'modified',
           );
           if (status === 'complete') {
-            dispatch(LocalInventoryActions.updatePendingCount('decrement'));
-            dispatch(LocalInventoryActions.updateUploadCount('decrement'));
+            updateCount({ type: 'pending', count: 'decrement' })(dispatch);
+            updateCount({ type: 'upload', count: 'decrement' })(dispatch);
           } else if (status === 'pending') {
-            dispatch(LocalInventoryActions.updatePendingCount('increment'));
+            updateCount({ type: 'pending', count: 'increment' })(dispatch);
           }
           resolve();
         });
@@ -121,10 +121,10 @@ export const changeInventoryStatus = ({ inventory_id, status }, dispatch) => {
             'modified',
           );
           if (status === 'complete') {
-            dispatch(LocalInventoryActions.updatePendingCount('decrement'));
-            dispatch(LocalInventoryActions.updateUploadCount('decrement'));
+            updateCount({ type: 'pending', count: 'decrement' })(dispatch);
+            updateCount({ type: 'upload', count: 'decrement' })(dispatch);
           } else if (status === 'pending') {
-            dispatch(LocalInventoryActions.updatePendingCount('increment'));
+            updateCount({ type: 'pending', count: 'increment' })(dispatch);
           }
           resolve();
         });
@@ -167,5 +167,374 @@ export const updateSpecieDiameter = ({ inventory_id, speciesDiameter }) => {
         resolve();
       })
       .catch(bugsnag.notify);
+  });
+};
+
+export const deleteInventory = ({ inventory_id }, dispatch) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', `${inventory_id}`);
+          realm.delete(inventory);
+        });
+        if (dispatch) {
+          updateCount({ type: 'pending', count: 'decrement' })(dispatch);
+        }
+        resolve(true);
+      })
+      .catch((err) => {
+        reject(err);
+        bugsnag.notify(err);
+      });
+  });
+};
+
+export const updatePlantingDate = ({ inventory_id, plantation_date }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          realm.create(
+            'Inventory',
+            {
+              inventory_id: `${inventory_id}`,
+              plantation_date,
+            },
+            'modified',
+          );
+          resolve();
+        });
+      })
+      .catch(bugsnag.notify);
+  });
+};
+
+export const updateLastScreen = ({ last_screen, inventory_id }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          realm.create(
+            'Inventory',
+            {
+              inventory_id: `${inventory_id}`,
+              last_screen: last_screen,
+            },
+            'modified',
+          );
+          resolve();
+        });
+      })
+      .catch(bugsnag.notify);
+  });
+};
+
+export const clearAllIncompleteInventory = () => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let allInventory = realm.objects('Inventory').filtered('status == "incomplete"');
+          realm.delete(allInventory);
+          resolve();
+        });
+      })
+      .catch(bugsnag.notify);
+  });
+};
+
+export const clearAllUploadedInventory = () => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let allInventory = realm.objects('Inventory').filtered('status == "complete"');
+          realm.delete(allInventory);
+          resolve();
+        });
+      })
+      .catch(bugsnag.notify);
+  });
+};
+
+export const updateSpecieAndSpecieDiameter = ({ inventory_id, specie_name, diameter }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', `${inventory_id}`);
+          inventory.species_diameter = Number(diameter);
+          inventory.specei_name = specie_name;
+        });
+        resolve();
+      })
+      .catch((err) => {
+        reject(err);
+        bugsnag.notify(err);
+      });
+  });
+};
+
+export const removeLastCoord = ({ inventory_id }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', `${inventory_id}`);
+          let polygons = Object.values(JSON.parse(JSON.stringify(inventory.polygons)));
+          let coords = Object.values(polygons[polygons.length - 1].coordinates);
+          coords = coords.slice(0, coords.length - 1);
+          polygons[polygons.length - 1].coordinates = coords;
+          inventory.polygons = polygons;
+          resolve();
+        });
+      })
+      .catch(bugsnag.notify);
+  });
+};
+
+export const getCoordByIndex = ({ inventory_id, index }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', `${inventory_id}`);
+          let polygons = Object.values(JSON.parse(JSON.stringify(inventory.polygons)));
+          let coords = Object.values(polygons[0].coordinates);
+          let coordsLength = coords.length;
+          resolve({ coordsLength, coord: coords[index] });
+        });
+      })
+      .catch(bugsnag.notify);
+  });
+};
+
+export const insertImageAtIndexCoordinate = ({ inventory_id, imageUrl, index }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', `${inventory_id}`);
+          let polygons = Object.values(JSON.parse(JSON.stringify(inventory.polygons)));
+          let polygonsTemp = [];
+          let coordinatesTemp = [];
+
+          polygonsTemp = polygons.map((onePolygon, i) => {
+            let coords = Object.values(onePolygon.coordinates);
+            coords[index].imageUrl = imageUrl;
+            return { isPolygonComplete: onePolygon.isPolygonComplete, coordinates: coords };
+          });
+          inventory.polygons = polygonsTemp;
+
+          resolve();
+        });
+      })
+      .catch(bugsnag.notify);
+  });
+};
+
+export const addCoordinates = ({ inventory_id, geoJSON, currentCoords }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let polygons = [];
+          geoJSON.features.map((onePolygon) => {
+            let onePolygonTemp = {};
+            onePolygonTemp.isPolygonComplete = onePolygon.properties.isPolygonComplete || false;
+            let coordinates = [];
+            onePolygon.geometry.coordinates.map((oneLatlong) => {
+              coordinates.push({
+                longitude: oneLatlong[0],
+                latitude: oneLatlong[1],
+                currentloclat: currentCoords.latitude ? currentCoords.latitude : 0,
+                currentloclong: currentCoords.longitude ? currentCoords.longitude : 0,
+                isImageUploaded: false,
+              });
+            });
+            onePolygonTemp.coordinates = coordinates;
+            polygons.push(onePolygonTemp);
+          });
+          realm.create(
+            'Inventory',
+            {
+              inventory_id: `${inventory_id}`,
+              polygons: polygons,
+            },
+            'modified',
+          );
+          resolve();
+        });
+      })
+      .catch((err) => {
+        reject(err);
+        bugsnag.notify(err);
+      });
+  });
+};
+
+export const addCoordinateSingleRegisterTree = ({
+  inventory_id,
+  markedCoords,
+  currentCoords,
+  locateTree,
+}) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', inventory_id);
+          inventory.polygons = [
+            {
+              coordinates: [
+                {
+                  latitude: markedCoords[1],
+                  longitude: markedCoords[0],
+                  currentloclat: currentCoords.latitude,
+                  currentloclong: currentCoords.longitude,
+                  isImageUploaded: false,
+                },
+              ],
+            },
+          ];
+          inventory.species_diameter = 10;
+          locateTree ? (inventory.locate_tree = locateTree) : null;
+          inventory.plantation_date = `${Date.now()}`;
+          resolve();
+        });
+      })
+      .catch((err) => {
+        reject(err);
+        bugsnag.notify(err);
+      });
+  });
+};
+
+export const insertImageSingleRegisterTree = ({ inventory_id, imageUrl }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', inventory_id);
+          inventory.polygons[0].coordinates[0].imageUrl = imageUrl;
+          resolve();
+        });
+      })
+      .catch((err) => {
+        reject(err);
+        bugsnag.notify(err);
+      });
+  });
+};
+
+export const addSpeciesAction = ({ inventory_id, species, plantation_date }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          realm.create(
+            'Inventory',
+            {
+              inventory_id: `${inventory_id}`,
+              species,
+              plantation_date,
+            },
+            'modified',
+          );
+          resolve();
+        });
+      })
+      .catch((err) => {
+        reject(err);
+        bugsnag.notify(err);
+      });
+  });
+};
+
+export const addLocateTree = ({ locate_tree, inventory_id }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          realm.create(
+            'Inventory',
+            {
+              inventory_id: `${inventory_id}`,
+              locate_tree: locate_tree,
+            },
+            'modified',
+          );
+          resolve();
+        });
+      })
+      .catch(bugsnag.notify);
+  });
+};
+
+export const polygonUpdate = ({ inventory_id }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', `${inventory_id}`);
+          inventory.polygons[0].isPolygonComplete = true;
+          resolve();
+        });
+      })
+      .catch((err) => {
+        reject(err);
+        bugsnag.notify(err);
+      });
+  });
+};
+
+export const completePolygon = ({ inventory_id }) => {
+  return new Promise((resolve, reject) => {
+    Realm.open({
+      schema: [Inventory, Species, Polygons, Coordinates, OfflineMaps, User, AddSpecies],
+    })
+      .then((realm) => {
+        realm.write(() => {
+          let inventory = realm.objectForPrimaryKey('Inventory', `${inventory_id}`);
+          inventory.polygons[0].isPolygonComplete = true;
+          inventory.polygons[0].coordinates.push(inventory.polygons[0].coordinates[0]);
+          resolve();
+        });
+      })
+      .catch((err) => {
+        reject(err);
+        bugsnag.notify(err);
+      });
   });
 };
